requirements.txt
# Core ML & Data
pandas>=2.3.0
scipy>=1.13.1
scikit-learn>=1.6.1
xgboost>=2.1.4
shap>=0.47.2

# Orchestration
dagster[postgresql]>=1.10.19
feast>=0.49.0
# Note: Pulser CLI v2.2.2 is managed via npm link, not pip

# AI & Vector Search
openai>=1.84.0
pinecone-client>=6.0.0

# API & Web
fastapi>=0.115.12
uvicorn>=0.34.0
pydantic>=2.10.6
python-dotenv>=1.0.1

# Azure
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0

# Utils
joblib>=1.5.1

sqlalchemy==2.0.28
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
prefect>=2.0.0
rclone-python>=0.1.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0 